+++ bindings/python/IK/src/SE3Task.cpp
@@ -11,6 +11,8 @@
 
 #include <BipedalLocomotion/IK/IKLinearTask.h>
 #include <BipedalLocomotion/IK/SE3Task.h>
+#include <BipedalLocomotion/System/ITaskControllerManager.h>
+
 #include <BipedalLocomotion/bindings/IK/SE3Task.h>
 
 namespace BipedalLocomotion
@@ -25,7 +27,10 @@ void CreateSE3Task(pybind11::module& module)
     namespace py = ::pybind11;
     using namespace BipedalLocomotion::IK;
 
-    py::class_<SE3Task, std::shared_ptr<SE3Task>, IKLinearTask>(module, "SE3Task")
+    py::class_<SE3Task,
+               std::shared_ptr<SE3Task>,
+               IKLinearTask,
+               BipedalLocomotion::System::ITaskControllerManager>(module, "SE3Task")
         .def(py::init())
         .def("set_kin_dyn", &SE3Task::setKinDyn, py::arg("kin_dyn"))
         .def("set_set_point", &SE3Task::setSetPoint, py::arg("I_H_F"), py::arg("mixed_velocity"));