+++ bindings/python/TSID/src/SE3Task.cpp
@@ -9,8 +9,10 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include <BipedalLocomotion/TSID/TSIDLinearTask.h>
+#include <BipedalLocomotion/System/ITaskControllerManager.h>
 #include <BipedalLocomotion/TSID/SE3Task.h>
+#include <BipedalLocomotion/TSID/TSIDLinearTask.h>
+
 #include <BipedalLocomotion/bindings/TSID/SE3Task.h>
 
 namespace BipedalLocomotion
@@ -25,7 +27,8 @@ void CreateSE3Task(pybind11::module& module)
     namespace py = ::pybind11;
     using namespace BipedalLocomotion::TSID;
 
-    py::class_<SE3Task, std::shared_ptr<SE3Task>, TSIDLinearTask>(module, "SE3Task")
+    py::class_<SE3Task, std::shared_ptr<SE3Task>, TSIDLinearTask,
+	       BipedalLocomotion::System::ITaskControllerManager>(module, "SE3Task")
         .def(py::init())
         .def("set_kin_dyn", &SE3Task::setKinDyn, py::arg("kin_dyn"))
         .def("set_set_point",